Add static file support
By having a public directory, those files will be served statically,
automatically.

An example is placed in the examples directory; this means that if you
want to try an example with `cargo run --example`, you have to be in the
examples directory rather than the root to see the static files.