Ensure we always send a Content-Length header
ie. for every response, not just the ones where the user callback
returned Ok.

This changes the previous behaviour in that Ok case, since we now
only write this header for the consumer if they don't set it on
their own. This means the consumer is now allowed to lie and send
the wrong length.

write_response was reflowed to make it easier to conditionally
insert something into the text. As a bonus, it is now slightly
less allocation-heavy.

This also removes the Cow machinery that isn't used since Server
isn't generic anymore.